+++ libs/expo/betterangels/src/lib/screenRouting/clientProfileRoutes/getEditClientProfileRoute.ts
@@ -1,0 +1,17 @@
+import { TStandardSection } from './types';
+
+type TGetRouteProps = {
+  profileId: string;
+  section: TStandardSection;
+};
+
+export function getEditClientProfileRoute(props: TGetRouteProps) {
+  const { profileId, section } = props;
+
+  return {
+    pathname: `/clients/${profileId}/edit`,
+    params: {
+      componentName: section,
+    },
+  };
+}
