+++ libs/expo/betterangels/src/lib/screenRouting/clientProfileRoutes/index.ts
@@ -2,11 +2,10 @@ export {
   ClientProfileSectionEnum,
   isValidClientProfileSectionEnum,
 } from './constants';
-export { getClientProfileRoute } from './getClientProfileRoute';
 export {
-  getClientProfileEditRoute,
   getRelatedModelAddRoute,
   getRelatedModelEditRoute,
-  getRelatedModelsViewRoute,
-} from './getClientProfileScreenRoute';
+  getRelatedModelViewRoute,
+} from './getClientProfileRelatedModelRoute';
+export { getEditClientProfileRoute } from './getEditClientProfileRoute';
 export { TRelatedModelSection, TStandardSection } from './types';
